First work on search page
Added a new sproc to get search terms, and a route to fetch them. New code in the search page pulls these in and updates the form with terms from the DB.

Cleaned up CSS so the site is maringally less ugly.

This checkin also removes the dependency on the old site hierarchy (which means all of the pictures go away, temporarily).